fix clang warning
git-svn-id: svn://svn.code.sf.net/p/lush/code@2813 0babfeb2-de4b-461b-a2bf-1ab085750f0d

+++ lush1/trunk/src/storage.c
@@ -1006,7 +1006,7 @@ storage_mmap(at *atp, FILE *f, size_t offset)
   st->allinfo.sts_mmap.xtra = xtra;
   st->allinfo.sts_mmap.addr = addr;
   st->srg.data = (char*)(st->allinfo.sts_mmap.addr)+offset;
-  st->srg.flags = (STS_MMAP | STF_RDONLY) & ~STF_UNSIZED;
+  st->srg.flags = (short)((STS_MMAP | STF_RDONLY) & ~STF_UNSIZED);
 }
 
 DX(xstorage_mmap)
